Rename os support modules
